fix: rename useScrollDisabled to useScrollLock
