--- packages/hooks/docs/hooks/useScrollDisabled.story.mdx
@@ -1,46 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import { WrapperScrollDisabled } from '../stories';
-
-<Meta title="Modules|Hooks/useScrollDisabled" />
-
-# Overview
-
-`useScrollDisabled` A hook that elegantly scroll disabled.
-
-### API
-
-```ts
-    useScrollDisabled(
-        disabled: boolean
-    );
-```
-
-### Params
-
-| Property | Description                  | Type    | Default |
-| -------- | ---------------------------- | ------- | ------- |
-| disabled | Enable or disable the scroll | boolean | true    |
-
-### Usage
-
-```js
-import React, { useState } from 'react';
-import { CheckboxToggle } from 'react-rainbow-components';
-import { useScrollDisabled } from '../../src';
-
-export default function WrapperScrollDisabled() {
-    const [isDisabled, setIsDisabled] = useState(false);
-    useScrollDisabled(isDisabled);
-    return (
-        <CheckboxToggle
-            label="Enable and disabled scroll"
-            value={isDisabled}
-            onChange={(event) => setIsDisabled(event.target.checked)}
-        />
-    );
-}
-```
-
-### Examples
-
-<WrapperScrollDisabled />
