+++ packages/hooks/docs/stories/index.js
@@ -1,3 +1,3 @@
 export { default as WrapperOutsideClick } from './wrapperOutsideClick';
-export { default as WrapperScrollDisabled } from './wrapperScrollDisabled';
+export { default as WrapperScrollLock } from './wrapperScrollLock';
 export { default as WrapperUniqueIdentifier } from './wrapperUniqueIdentifier';
