+++ packages/hooks/src/index.js
@@ -1,3 +1,3 @@
 export { default as useOutsideClick } from './useOutsideClick';
-export { default as useScrollDisabled } from './useScrollDisabled';
+export { default as useScrollLock } from './useScrollLock';
 export { default as useUniqueIdentifier } from './useUniqueIdentifier';
